Fixed rook move errors, finished move generation routine in gen.py

+++ move.py
@@ -3,24 +3,24 @@ from numpy import uint64
 class Move(uint64):
   @property
   def from_(self):
-    return self & 63
+    return self & uint64(63)
 
   @property
   def to(self):
-   return  self >> 6 & 63
+   return  self >> uint64(6) & uint64(63)
 
   @property
   def piece(self):
-    return self >> 12 & 7
+    return self >> uint64(12) & uint64(7)
 
   @property
   def captured(self):
-    return self >> 15 & 7
+    return self >> uint64(15) & uint64(7)
 
   @property
   def promoted(self):
-    return self >> 18 & 7
+    return self >> uint64(18) & uint64(7)
 
   @property
   def flip(self):
-    return self ^ 1
+    return self ^ uint64(1)
